# Try Adaptive Radix Tree for index