Just to see how it'd look vs map. Ref: https://github.com/plar/go-adaptive-radix-tree