@eiri 2019-04-04
## Bench

```
benchcmp before.bench after.bench
benchmark                old ns/op     new ns/op     delta
BenchmarkIndexLoad-4     78610014      179395067     +128.21%
BenchmarkOpen-4          77859736      182519501     +134.42%
BenchmarkWrite-4         121408        126466        +4.17%
BenchmarkRead-4          2100          2156          +2.67%

benchmark                old allocs     new allocs     delta
BenchmarkIndexLoad-4     9579           1338274        +13870.92%
BenchmarkOpen-4          9637           1338294        +13787.04%
BenchmarkWrite-4         1              7              +600.00%
BenchmarkRead-4          1              2              +100.00%

benchmark                old bytes     new bytes     delta
BenchmarkIndexLoad-4     41060316      36570225      -10.94%
BenchmarkOpen-4          41077187      36571680      -10.97%
BenchmarkWrite-4         1280          1313          +2.58%
BenchmarkRead-4          16            32            +100.00%
```