@eiri 2019-04-04
### Summary
Well, nothing surprising here. Slightly slower read/write with overall noticeably better memory footprint. Nice and easy interface in the lib itself. Would be a way to go if I'd need range queries or prefix lookup, but it'd be wasted on the random score's nature.

### References
Original build: https://travis-ci.com/eiri/dieci/builds/102258187